+++ src/views/DataTable/LoaderRow.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import SpinnerLoader from '../SpinnerLoader';
 
 const LoaderRow = ({ colSpan }) => (
@@ -8,5 +9,9 @@ const LoaderRow = ({ colSpan }) => (
         </td>
     </tr>
 );
+
+LoaderRow.propTypes =  {
+    colSpan: PropTypes.number
+};
 
 export default LoaderRow;