+++ src/views/LoadingAlert.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Alert } from 'reactstrap';
 
 const LoadingAlert = ({ msg }) => (
@@ -8,5 +9,9 @@ const LoadingAlert = ({ msg }) => (
         {msg}
     </Alert>
 );
+
+LoadingAlert.propTypes = {
+    msg: PropTypes.string
+};
 
 export default LoadingAlert;